requirements.txt
streamlit>=1.24
sentence-transformers>=2.2.2
# faiss-cpu on Windows can be tricky; if installation fails, see README for alternatives
faiss-cpu>=1.7.4
numpy>=1.23
pandas>=1.5
tqdm>=4.64
pdfminer.six>=20221105
cross-encoder>=0.4.3
pyarrow>=8.0.0
